requirements.txt
python-dotenv
pandas
openai
openpyxl